# ♻️Use singleton instances of attributes