--- a/src/AutoFilterer/Attributes/ArraySearchFilterAttribute.cs
+++ b/src/AutoFilterer/Attributes/ArraySearchFilterAttribute.cs
@@ -1,11 +1,7 @@
-﻿using System;
-using System.Collections;
-using System.Collections.Generic;
+﻿using System.Collections;
 using System.Linq;
 using System.Linq.Expressions;
 using System.Reflection;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace AutoFilterer.Attributes;